Adds league_stats class and test
