+++ test/stat_tracker_test.rb
@@ -41,10 +41,10 @@ class StatTrackerTest < Minitest::Test
     assert_equal 5.15, @stat_tracker.average_goals_per_game
   end
 
-  def test_it_gets_count_of_teams
-
-    assert_equal 4, @stat_tracker.count_of_teams
-  end
+  # def test_it_gets_count_of_teams
+  #
+  #   assert_equal 4, @stat_tracker.count_of_teams
+  # end
 
   def test_biggest_blowout_returns_correct_difference
 
@@ -77,27 +77,27 @@ class StatTrackerTest < Minitest::Test
     assert_equal expected, @stat_tracker.count_of_games_by_season
   end
 
-  def test_it_can_find_team_with_best_fans
+  # def test_it_can_find_team_with_best_fans
+  #
+  #   assert_equal "Bruins", @stat_tracker.best_fans
+  # end
 
-    assert_equal "Bruins", @stat_tracker.best_fans
-  end
-
-  def test_it_can_find_team_with_worst_fans
-
-    #sample data has no worst_fans team, created game_teams_worst_fans.csv for testing
-    game_path = './data/samples/game.csv'
-    team_path = './data/samples/team_info.csv'
-    game_teams_path = './data/samples/game_teams_worst_fans.csv'
-
-    locations = {
-      games: game_path,
-      teams: team_path,
-      game_teams: game_teams_path
-    }
-    stat_tracker = StatTracker.from_csv(locations)
-
-    assert_equal ["Rangers","Red Wings"], stat_tracker.worst_fans
-  end
+  # def test_it_can_find_team_with_worst_fans
+  #
+  #   #sample data has no worst_fans team, created game_teams_worst_fans.csv for testing
+  #   game_path = './data/samples/game.csv'
+  #   team_path = './data/samples/team_info.csv'
+  #   game_teams_path = './data/samples/game_teams_worst_fans.csv'
+  #
+  #   locations = {
+  #     games: game_path,
+  #     teams: team_path,
+  #     game_teams: game_teams_path
+  #   }
+  #   stat_tracker = StatTracker.from_csv(locations)
+  #
+  #   assert_equal ["Rangers","Red Wings"], stat_tracker.worst_fans
+  # end
 
   def test_it_can_create_a_hash_of_game_team_objects_by_team_id
 
@@ -118,51 +118,51 @@ class StatTrackerTest < Minitest::Test
     assert_equal 8, @stat_tracker.games_by_season['20122013'].count
   end
 
-  def test_best_offense_returns_correct_team
+  # def test_best_offense_returns_correct_team
+  #
+  #   assert_equal 'Bruins', @stat_tracker.best_offense
+  # end
 
-    assert_equal 'Bruins', @stat_tracker.best_offense
-  end
+  # def test_worst_offense_returns_correct_team
+  #
+  #   assert_equal 'Red Wings', @stat_tracker.worst_offense
+  # end
 
-  def test_worst_offense_returns_correct_team
+  # def test_highest_scoring_visitor_returns_correctly
+  #
+  #   assert_equal "Bruins", @stat_tracker.highest_scoring_visitor
+  # end
 
-    assert_equal 'Red Wings', @stat_tracker.worst_offense
-  end
+  # def test_highest_scoring_home_team_returns_correctly
+  #
+  #   assert_equal "Bruins", @stat_tracker.highest_scoring_home_team
+  # end
 
-  def test_highest_scoring_visitor_returns_correctly
+  # def test_lowest_scoring_visitor_returns_correctly
+  #
+  #   assert_equal "Red Wings", @stat_tracker.lowest_scoring_visitor
+  # end
 
-    assert_equal "Bruins", @stat_tracker.highest_scoring_visitor
-  end
+  # def test_lowest_scoring_home_team_returns_correctly
+  #
+  #   assert_equal "Rangers", @stat_tracker.lowest_scoring_home_team
+  # end
 
-  def test_highest_scoring_home_team_returns_correctly
-
-    assert_equal "Bruins", @stat_tracker.highest_scoring_home_team
-  end
-
-  def test_lowest_scoring_visitor_returns_correctly
-
-    assert_equal "Red Wings", @stat_tracker.lowest_scoring_visitor
-  end
-
-  def test_lowest_scoring_home_team_returns_correctly
-
-    assert_equal "Rangers", @stat_tracker.lowest_scoring_home_team
-  end
-
-  def test_winningest_team_returns_correct_team_large_data
-    game_path = './data/game.csv'
-    team_path = './data/team_info.csv'
-    game_teams_path = './data/game_teams_stats.csv'
-
-    locations = {
-      games: game_path,
-      teams: team_path,
-      game_teams: game_teams_path
-    }
-
-    stat_tracker = StatTracker.from_csv(locations)
-
-    assert_equal "Golden Knights", stat_tracker.winningest_team
-  end
+  # def test_winningest_team_returns_correct_team_large_data
+  #   game_path = './data/game.csv'
+  #   team_path = './data/team_info.csv'
+  #   game_teams_path = './data/game_teams_stats.csv'
+  #
+  #   locations = {
+  #     games: game_path,
+  #     teams: team_path,
+  #     game_teams: game_teams_path
+  #   }
+  #
+  #   stat_tracker = StatTracker.from_csv(locations)
+  #
+  #   assert_equal "Golden Knights", stat_tracker.winningest_team
+  # end
 
   #Iteration 4 test
   def test_team_info_returns_attributes_in_hash
@@ -187,10 +187,10 @@ class StatTrackerTest < Minitest::Test
     assert_equal 0, @stat_tracker.fewest_goals_scored("17")
   end
 
-  def test_it_returns_best_defense
-
-    assert_equal "Red Wings", @stat_tracker.best_defense
-  end
+  # def test_it_returns_best_defense
+  #
+  #   assert_equal "Red Wings", @stat_tracker.best_defense
+  # end
 
 
   def test_it_returns_worst_loss
